+++ src/remote_call/api/utils/ErrorUtils.h
@@ -32,150 +32,180 @@ enum class ErrorReason {
 class RemoteCallError : public ll::ErrorInfoBase {
     ErrorReason                      mReason;
     mutable std::string              mMessage;
-    mutable std::vector<std::string> mFiled{};
+    mutable std::vector<std::string> mFields{};
     std::optional<ll::Error>         mOriginError{};
 
 public:
-    inline RemoteCallError(ErrorReason reason, std::string&& msg, std::optional<ll::Error> originError = {}) noexcept
+    inline RemoteCallError(ErrorReason reason, std::string&& msg, std::optional<ll::Error> originError = {})
     : mReason(reason),
       mMessage(std::move(msg)),
       mOriginError(std::move(originError)) {}
+
     inline ~RemoteCallError() override = default;
+
     [[nodiscard]] constexpr std::string message() const noexcept override {
         if (!mOriginError) return mMessage;
         return fmt::format("{}\nOrigin Error: {}", mMessage, mOriginError->message());
     }
+
     constexpr RemoteCallError& append(std::string_view msg) {
-        if (!mMessage.ends_with("\n")) mMessage.append("\n");
-        mMessage.append(msg);
+        if (!mMessage.ends_with("\n")) mMessage += "\n";
+        mMessage += msg;
         return *this;
     }
+
     template <typename Fn = void>
     inline RemoteCallError& append(std::string_view ns, std::string_view func) {
-        if (!mMessage.ends_with("\n")) mMessage.append("\n");
-        auto  provider     = getProvider(ns, func).lock();
-        auto& providerName = provider ? provider->getName() : "UnknownMod";
-        mMessage.append(fmt::format(
+        if (!mMessage.ends_with("\n")) mMessage += "\n";
+        auto  provider      = getProvider(ns, func).lock();
+        auto& providerName  = provider ? provider->getName() : "UnknownMod";
+        mMessage           += fmt::format(
             "Function: [{}::{}](signature {}) provided by <{}>.\n",
             ns,
             func,
             type_raw_name_v<Fn>,
             providerName
-        ));
+        );
         return *this;
     }
+
     constexpr RemoteCallError& joinField(std::string_view field) {
-        mFiled.emplace_back(field);
+        mFields.emplace_back(field);
         return *this;
     }
+
     constexpr RemoteCallError& flushFields(std::string_view msg, std::string_view valueName) {
-        if (!mMessage.ends_with("\n")) mMessage.append("\n");
+        if (!mMessage.ends_with("\n")) mMessage += "\n";
         if (msg.ends_with("\n")) msg.remove_suffix(1);
-        std::string field = std::string(valueName);
-        for (auto&& f : mFiled | std::views::reverse) field.append(f);
-        mMessage.append(msg).append(" Field: ").append(field).append("\n");
-        mFiled.clear();
+        std::string fieldPath;
+        fieldPath.reserve(valueName.size() + mFields.size() * 10);
+
+        fieldPath += valueName;
+        for (auto& field : std::ranges::reverse_view(mFields)) {
+            fieldPath += field;
+        }
+        mMessage += fmt::format("{} Field: {}\n", msg, fieldPath);
+        mFields.clear();
         return *this;
     }
+
     [[nodiscard]] constexpr ErrorReason reason() const noexcept { return mReason; }
 
-    [[nodiscard]] LL_CONSTEXPR23 static bool isRemoteCallError(ll::Error& error) {
-        /// TODO:
-        return error.isA<RemoteCallError>();
+    [[nodiscard]] inline static bool isRemoteCallError(ll::Error& error) noexcept {
+        auto info = &error.as<ll::ErrorInfoBase>();
+        try {
+            return dynamic_cast<RemoteCallError*>(info) != nullptr;
+        } catch (...) {
+            return false;
+        }
     }
 };
 
+struct RemoteCallErrorWrapper {
+    ll::Error error;
 
-[[nodiscard]] LL_CONSTEXPR23 std::unique_ptr<RemoteCallError>
-                             convertUnknownError(std::string_view msg, ll::Error&& error) {
-    return std::make_unique<RemoteCallError>(
-        ErrorReason::Unknown,
-        fmt::format("{}: {}", msg, error.as<ll::ErrorInfoBase>().message()),
-        std::move(error)
-    );
-}
+    inline explicit RemoteCallErrorWrapper(ll::Error err, std::string_view unknownErrorMsg = "")
+    : error(
+          RemoteCallError::isRemoteCallError(err)
+              ? std::move(err)
+              : ll::Error{std::make_unique<RemoteCallError>(
+                    ErrorReason::Unknown,
+                    fmt::format("{}: {}", unknownErrorMsg, err.as<ll::ErrorInfoBase>().message()),
+                    std::move(err)
+                )}
+      ) {}
 
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected makeError(ErrorReason reason, std::string_view msg) {
-    auto error = std::make_unique<RemoteCallError>(reason, std::string(msg));
-    return ll::Unexpected(std::in_place, std::move(error));
+    inline explicit RemoteCallErrorWrapper(ErrorReason reason, std::string&& msg)
+    : error(std::make_unique<RemoteCallError>(reason, std::move(msg))) {}
+
+    RemoteCallError& get() noexcept { return error.as<RemoteCallError>(); }
+
+    inline operator ll::Unexpected() noexcept { return ll::forwardError(error); } // NOLINT(google-explicit-constructor)
+
+    [[nodiscard]] std::string message() noexcept { return get().message(); }
+    // clang-format off
+    constexpr RemoteCallErrorWrapper& append(std::string_view msg) { get().append(msg); return *this; }
+    template <typename Fn = void>
+    constexpr RemoteCallErrorWrapper& append(std::string_view ns, std::string_view func) { get().append<Fn>(ns, func); return *this; }
+    constexpr RemoteCallErrorWrapper& joinField(std::string_view field) { get().joinField(field); return *this; }
+    constexpr RemoteCallErrorWrapper& flushFields(std::string_view msg, std::string_view valueName) { get().flushFields(msg, valueName); return *this; }
+    [[nodiscard]] ErrorReason reason() noexcept { return get().reason(); }
+    // clang-format on
+};
+
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected makeError(ErrorReason reason, std::string_view msg) {
+    return RemoteCallErrorWrapper(reason, std::string(msg));
 }
 
 [[nodiscard]] LL_NOINLINE inline ll::Unexpected
 makeError(std::string_view ns, std::string_view func, ErrorReason reason, std::string_view msg) {
-    auto error = std::make_unique<RemoteCallError>(reason, std::string(msg));
-    error->append(ns, func);
-    return ll::Unexpected(std::in_place, std::move(error));
+    return RemoteCallErrorWrapper(reason, std::string{msg}) //
+        .append(ns, func);
 }
 
 template <typename Fn>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
 makeUnknownError(std::string_view ns, std::string_view func, std::string_view msg, ll::Error&& error) {
-    auto newError = convertUnknownError(msg, std::move(error));
-    newError->append<Fn>(ns, func);
-    return ll::Unexpected(std::in_place, std::move(newError));
+    return RemoteCallErrorWrapper(std::move(error), msg) //
+        .template append<Fn>(ns, func);
 }
 
 [[nodiscard]] LL_NOINLINE inline ll::Unexpected makeNotFoundError(std::string_view ns, std::string_view func) {
-    return makeError(ns, func, ErrorReason::NotExported, "Fail to import! Function has not been exported.");
+    return RemoteCallErrorWrapper(ErrorReason::NotExported, "Fail to import! Function has not been exported.")
+        .append(ns, func);
 }
-
 
 [[nodiscard]] LL_NOINLINE inline ll::Unexpected makeDisabledError(std::string_view ns, std::string_view func) {
-    return makeError(ns, func, ErrorReason::ProviderDisabled, "Fail to import! Provider has not been disabled.");
+    return RemoteCallErrorWrapper(ErrorReason::ProviderDisabled, "Fail to import! Provider has been disabled.")
+        .append(ns, func);
 }
 
 template <typename Ret, typename... Args>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
 makeArgsCountError(std::string_view ns, std::string_view func, size_t provided) {
-    auto error = std::make_unique<RemoteCallError>(
-        ErrorReason::ArgsCountNotMatch,
-        fmt::format("Fail to invoke! function requires {} args, but {} provided.", sizeof...(Args), provided)
-    );
-    error->append<Ret(Args...)>(ns, func);
-    return ll::Unexpected(std::in_place, std::move(error));
+    return RemoteCallErrorWrapper(
+               ErrorReason::ArgsCountNotMatch,
+               fmt::format("Fail to invoke! Function requires {} args, but {} provided.", sizeof...(Args), provided)
+    )
+        .template append<Ret(Args...)>(ns, func);
 }
 
 template <typename Ret, typename... Args>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
 makeSerializeError(std::string_view ns, std::string_view func, std::string_view valueName, ll::Error& error) {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        auto& newError = error.as<RemoteCallError>();
-        newError.flushFields("Failed to serialize value.", valueName);
-        newError.append<Ret(Args...)>(ns, func);
-        return ll::forwardError(error);
+    bool isRemoteCallError = RemoteCallError::isRemoteCallError(error);
+    auto wrapper           = RemoteCallErrorWrapper(std::move(error), "Serialization Error");
+    if (isRemoteCallError) {
+        wrapper.flushFields("Failed to serialize value", valueName);
     }
-    return makeUnknownError<Ret(Args...)>(ns, func, "Unknown Serialization Error: ", std::move(error));
+    return wrapper.template append<Ret(Args...)>(ns, func);
 }
 
 template <typename Ret, typename... Args>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
 makeDeserializeError(std::string_view ns, std::string_view func, std::string_view valueName, ll::Error& error) {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        auto& newError = error.as<RemoteCallError>();
-        newError.flushFields("Failed to deserialize value.", valueName);
-        newError.append<Ret(Args...)>(ns, func);
-        return ll::forwardError(error);
+    bool isRemoteCallError = RemoteCallError::isRemoteCallError(error);
+    auto wrapper           = RemoteCallErrorWrapper(std::move(error), "Deserialization Error");
+    if (isRemoteCallError) {
+        wrapper.flushFields("Failed to deserialize value", valueName);
     }
-    return makeUnknownError<Ret(Args...)>(ns, func, "Unknown Deserialization Error: {}", std::move(error));
+    return wrapper.template append<Ret(Args...)>(ns, func);
 }
 
 template <typename Ret, typename... Args>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
-                                         makeCallError(std::string_view ns, std::string_view func, ll::Error& error) {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        auto& newError = error.as<RemoteCallError>();
-        newError.append("Failed to call function!").append<Ret(Args...)>(ns, func);
-        return ll::forwardError(error);
-    }
-    return makeUnknownError<Ret(Args...)>(ns, func, "Unknown Call Error: {}", std::move(error));
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
+makeCallError(std::string_view ns, std::string_view func, ll::Error& error) {
+    return RemoteCallErrorWrapper(std::move(error), "Call Error")
+        .append("Failed to call function!")
+        .template append<Ret(Args...)>(ns, func);
 }
 
 template <typename Target, typename... Expected>
 [[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected makeFromDynamicTypeError(DynamicValue const& value) {
-    return makeError(
+    return RemoteCallErrorWrapper(
         ErrorReason::UnexpectedType,
         fmt::format(
-            "Failed to parse DynamicValue to {}. Expected alternative {}. Holding alternative {}.",
+            "Failed to convert DynamicValue to {}. Expected: [{}]. Actual: {}.",
             type_raw_name_v<Target>,
             typeListName<Expected...>(),
             typeName(value)
@@ -184,44 +214,33 @@ template <typename Target, typename... Expected>
 }
 
 template <typename T, typename Target>
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
-                                         makeUnsupportedValueError(std::string_view value, std::string_view msg) {
-    return makeError(
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected
+makeUnsupportedValueError(std::string_view value, std::string_view msg) {
+    return RemoteCallErrorWrapper(
         ErrorReason::UnsupportedValue,
-        fmt::format("Failed to convert {}<{}> to {}. {}", type_raw_name_v<T>, value, type_raw_name_v<Target>, msg)
+        fmt::format(
+            "Unsupported value conversion: {}<{}> -> {}. {}",
+            type_raw_name_v<T>,
+            value,
+            type_raw_name_v<Target>,
+            msg
+        )
     );
 }
 
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
-                                         makeSerMemberError(std::string_view name, ll::Error& error) noexcept {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        error.as<RemoteCallError>().joinField(fmt::format(".{}", name));
-        return ll::forwardError(error);
-    }
-    auto newError = convertUnknownError("Unknown Serialization Member Error", std::move(error));
-    newError->joinField(fmt::format(".{}", name));
-    return ll::Unexpected(std::in_place, std::move(newError));
-}
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected makeSerIndexError(std::size_t idx, ll::Error& error) noexcept {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        error.as<RemoteCallError>().joinField(fmt::format("[{}]", idx));
-        return ll::forwardError(error);
-    }
-    auto newError = convertUnknownError("Unknown Serialization Index Error", std::move(error));
-    newError->joinField(fmt::format("[{}]", idx));
-    return ll::Unexpected(std::in_place, std::move(newError));
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected makeSerMemberError(std::string_view name, ll::Error& error) {
+    return RemoteCallErrorWrapper(std::move(error), "Member Serialization Error") //
+        .joinField(fmt::format(".{}", name));
 }
 
-[[nodiscard]] LL_NOINLINE LL_CONSTEXPR23 ll::Unexpected
-                                         makeSerKeyError(std::string_view key, ll::Error& error) noexcept {
-    if (RemoteCallError::isRemoteCallError(error)) {
-        error.as<RemoteCallError>().joinField(fmt::format("[\"{}\"]", key));
-        return ll::forwardError(error);
-    }
-    auto newError = convertUnknownError("Unknown Serialization Key Error", std::move(error));
-    newError->joinField(fmt::format("[\"{}\"]", key));
-    return ll::Unexpected(std::in_place, std::move(newError));
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected makeSerIndexError(size_t idx, ll::Error& error) {
+    return RemoteCallErrorWrapper(std::move(error), "Index Serialization Error") //
+        .joinField(fmt::format("[{}]", idx));
 }
 
+[[nodiscard]] LL_NOINLINE inline ll::Unexpected makeSerKeyError(std::string_view key, ll::Error& error) {
+    return RemoteCallErrorWrapper(std::move(error), "Key Serialization Error") //
+        .joinField(fmt::format("[\"{}\"]", key));
+}
 
 } // namespace remote_call::error_utils